+++ golikov.george/A1/circle.hpp
@@ -6,12 +6,12 @@
 class Circle : public Shape
 {
   public:
-    float m_radius;
+    double m_radius;
     point_t m_center;
     double getArea() const override;
     rectangle_t getFrameRect() const override;
     void move(const point_t &pos) override;
-    void move(const float dltX, const float dltY) override;
+    void move(const double dltX, const double dltY) override;
 };
 
 #endif
